ADV 159
In this project, you will be given a GUI based application which displays an image of the flag of that country when the user enters the country name, but gives an error when entered a country name whose flag is not added to the application. You have to handle the error caused due to user input using try-except blocks.